Move WhisperMode to its own file.  Will probably eventually become a Whispers file containing all classes and functions pertaining to whispers and whisper management.
