+++ model/__init__.py
@@ -1,3 +1,4 @@
+from .whisper_mode import WhisperMode
 from .script import Script
 
-__all__ = ["Script"]
+__all__ = ["Script", "WhisperMode"]
